requirements.txt
Flask
pdfplumber
python-pptx
reportlab
transformers
language-tool-python
torch